refactor: move actuators services into own folder
- Prepare for Sensor Implementation

issue #43
