+++ app/services/actuators/proportional.py
@@ -1,12 +1,17 @@
 from app.models.actuators import ProportionalValve
+from app.services.actuators.service import ActuatorService
 from app.utils.config import settings
-from app.services.actuators import ActuatorService
+
 
 match settings.PROPORTIONAL_MODE:
     case "GPIO":
-        from app.repositories.GPIO.proportional_valves import ProportionalActuator
+        from app.repositories.actuators.GPIO.proportional_valves import (
+            ProportionalActuator,
+        )
     case "CAN":
-        from app.repositories.CAN.proportional_valves import ProportionalActuator
+        from app.repositories.actuators.CAN.proportional_valves import (
+            ProportionalActuator,
+        )
 
 
 class ProportionalService(ActuatorService[ProportionalValve]):
